Reformatted methods; Removed unused methods that wouldn't fully support context menus, and are not part of this core feature; Relocated view-util.service.ts to app/services
