+++ src/app/app.module.ts
@@ -84,7 +84,7 @@ import { NodePermissionsDialogComponent } from './dialogs/node-permissions/node-
 import { NodePermissionsDirective } from './directives/node-permissions.directive';
 import { PermissionsManagerComponent } from './components/permission-manager/permissions-manager.component';
 import { AppRouteReuseStrategy } from './app.routes.strategy';
-import { ViewUtilService} from './common/services/view-util.service';
+import { ViewUtilService} from './services/view-util.service';
 
 @NgModule({
     imports: [
