+++ src/app/components/files/files.component.ts
@@ -36,7 +36,6 @@ import { PageComponent } from '../page.component';
 import { ContentApiService } from '../../services/content-api.service';
 import { ExtensionService } from '../../extensions/extension.service';
 import { SetCurrentFolderAction } from '../../store/actions';
-import { ViewUtilService } from '../../common/services/view-util.service';
 
 @Component({
     templateUrl: './files.component.html'
@@ -53,7 +52,6 @@ export class FilesComponent extends PageComponent implements OnInit, OnDestroy {
                 store: Store<AppStore>,
                 private nodeActionsService: NodeActionsService,
                 private uploadService: UploadService,
-                private viewUtil: ViewUtilService,
                 content: ContentManagementService,
                 extensions: ExtensionService) {
         super(store, extensions, content);
@@ -250,14 +248,6 @@ export class FilesComponent extends PageComponent implements OnInit, OnDestroy {
                 elements.splice(docLib, 1);
             }
         }
-    }
-
-    /**
-     * Use this method to print files from a
-     * @param event
-     */
-    onPrintFile(event: any) {
-        this.viewUtil.printFileGeneric(event.value.entry.id, event.value.entry.content.mimeType);
     }
 
     isSiteContainer(node: MinimalNodeEntryEntity): boolean {
